write method & error handle add

+++ mixi_convart.rb
@@ -23,19 +23,33 @@ class Mixi_to_markdown
 		bodies = body
 	end
 
-	def make_markdown
-		doc = REXML::Document.new(open(Mixi_xml))
-		doc.elements.each("/diary/day") do |day|
-			title = day.attributes['title']
-			date = day.attributes['date']
-			body = day.elements['body'].text
-			#escape shash etc
-			title = title.gsub(/\//){|word|word.slice!(1)}
-			#実行部（別のメソッドにしたほうがいいのかな・・）
-			file_path = Post_path + make_filenames(date,title)
-			write_body = make_header(title,date) + make_bodies(body)
-
+	def file_write(file_path,write_body)
+		if (File.exist?(file_path))
+			puts "#{file_path} already exist"
+		else
+			puts Time.now.to_s + file_path + "作成しました。"
 			File.write(file_path,write_body)
+		end
+	end
+
+	def make_markdown
+		begin
+			doc = REXML::Document.new(open(Mixi_xml))
+			doc.elements.each("/diary/day") do |day|
+				title = day.attributes['title']
+				date = day.attributes['date']
+				body = day.elements['body'].text
+				#escape shash etc
+				title = title.gsub(/\//){|word|word.slice!(1)}
+				#実行部（別のメソッドにしたほうがいいのかな・・）
+				file_path = Post_path + make_filenames(date,title)
+				write_body = make_header(title,date) + make_bodies(body)
+
+				file_write(file_path,write_body)
+			end
+		rescue => ex
+			puts "読込先のXMLファイルが存在しないか、なんかのバグです"
+			puts "詳細：" + ex.message
 		end
 	end
 
@@ -43,4 +57,3 @@ end
 
 mixi = Mixi_to_markdown.new
 mixi.make_markdown
-puts 'ls -l /Users/hitokui/Lab/dory-dd.github.io/_posts/'
